removed labrador
