+++ extra/build.sbt
@@ -7,7 +7,6 @@ libraryDependencies ++= {
   Seq(
     "org.scalatest" %% "scalatest" % "3.0.5" % "test",
     "ai.lum"     %% "nxmlreader"            % "0.1.2",
-    "ai.lum"     %% "labrador-core"         % "0.0.2-SNAPSHOT",
     "org.clulab" %% "processors-main"       % procVersion,
     "org.clulab" %% "processors-modelsmain" % procVersion
   )
